NEW ARM LAUNCHING WOO HOO, need to change some stuff to get it working I think but idk what yet. Joint state publisher gui isn't quite qorking yet
