+++ sw_arm_desc_old/setup.cfg
@@ -1,0 +1,4 @@
+[develop]
+script_dir=$base/lib/sw_arm_desc_old
+[install]
+install_scripts=$base/lib/sw_arm_desc_old
