--- test_urdf_tool/package.xml
@@ -1,24 +1,0 @@
-<?xml version="1.0"?>
-<?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
-<package format="3">
-  <name>test_urdf_tool</name>
-  <version>0.0.0</version>
-  <description>TODO: Package description</description>
-  <maintainer email="olmer@gmail.com">ros-industrial</maintainer>
-  <license>TODO: License declaration</license>
-
-  <test_depend>ament_copyright</test_depend>
-  <test_depend>ament_flake8</test_depend>
-  <test_depend>ament_pep257</test_depend>
-  <test_depend>python3-pytest</test_depend>
-  
-        <depend>urdf</depend>
-      <build_depend>launch_ros</build_depend>
-      <exec_depend>launch_ros</exec_depend>
-      <exec_depend>robot_state_publisher</exec_depend>
-      <exec_depend>joint_state_publisher_gui</exec_depend>
-
-  <export>
-    <build_type>ament_python</build_type>
-  </export>
-</package>
