--- test_urdf_tool/setup.cfg
@@ -1,4 +1,0 @@
-[develop]
-script_dir=$base/lib/test_urdf_tool
-[install]
-install_scripts=$base/lib/test_urdf_tool
